% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_workflow_script.R
\name{add_script}
\alias{add_script}
\title{Add a new script to folder}
\usage{
add_script(name, wd = getwd(), renumber = FALSE)
}
\arguments{
\item{name}{String with name of the new file to create.}

\item{wd}{Path where to create new R script, defaults to working directory.}

\item{renumber}{Boolean. Default is FALSE. If the script name has a number which creates conflicts with existing scripts, you will need to set this parameter o TRUE in order to rename all necessary files.}
}
\description{
Add a new script to folder
}
