% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamer.R
\name{sha1}
\alias{sha1}
\title{Rename Files to Their SHA-1 Checksum}
\usage{
sha1(path = ".")
}
\arguments{
\item{path}{A character string specifying the path to the directory
containing the files. Defaults to the current working directory.}
}
\value{
None
}
\description{
This function renames all files in the specified directory to their
SHA-1 checksum followed by their original file extension. It only
processes files with the extensions `.jpg`, `.jpeg`, and `.png`.
}
\examples{
\dontrun{
rename("/path/to/your/directory")
}
}
