% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{under_score}
\alias{camelCase}
\alias{camelCase.}
\alias{under_score}
\title{The alias creators}
\usage{
under_score(package, package.lib = NULL)

camelCase(package, package.lib = NULL)
}
\arguments{
\item{package}{name (as a character string)}

\item{package.lib}{character vector specifying library. The default \code{NULL}, will
search through \code{.libPaths()} for location of package}
}
\description{
This package scans the list of exported functions and creates either
underscored or camelCase variants. For example, \sQuote{plotLike} becomes \sQuote{plot_lik}
and \sQuote{geom_point} becomes \sQuote{geomPoint}
}

