% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{underscore}
\alias{underscore}
\title{Detect and create underscored aliases}
\usage{
underscore(package, package.lib = NULL)
}
\arguments{
\item{package}{name (as a character string)}

\item{package.lib}{character vector specifying library. The default \code{NULL}, will
search through \code{.libPaths()} for location of package.}
}
\description{
This package scans the list of exported functions and creates
underscored variants. For example, \sQuote{plotLike} becomes \sQuote{plot_lik}
}

